pyproject.toml
[project]
name = "uma-prediction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "datasets>=4.0.0",
    "html5lib>=1.1",
    "huggingface-hub>=0.33.2",
    "lightgbm>=4.6.0",
    "lxml>=6.0.0",
    "numpy>2.0.0",
    "pandas>=2.3.0",
    "pyarrow>=20.0.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "scikit-learn>=1.7.0",
    "streamlit>=1.46.1",
    "tensorflow>=2.19.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
